Conflicts:
	sql/pitertul

Former-commit-id: 5ec5f08dcfc58df5a2dbd6ccb90ed0e653867454

+++ recovery-webservice/src/main/java/es/pfsgroup/plugin/rem/model/DtoActivosExpediente.java
@@ -11,7 +11,7 @@ public class DtoActivosExpediente extends WebDto {
 	private String subtipoActivo;
 	private Double precioMinimo;
 	private Double precioAprobadoVenta;
-	private Float porcentajeParticipacion;
+	private Double porcentajeParticipacion;
 	private Double importeParticipacion;
 	
 
@@ -63,11 +63,11 @@ public class DtoActivosExpediente extends WebDto {
 		this.precioAprobadoVenta = precioAprobadoVenta;
 	}
 
-	public Float getPorcentajeParticipacion() {
+	public Double getPorcentajeParticipacion() {
 		return porcentajeParticipacion;
 	}
 
-	public void setPorcentajeParticipacion(Float porcentajeParticipacion) {
+	public void setPorcentajeParticipacion(Double porcentajeParticipacion) {
 		this.porcentajeParticipacion = porcentajeParticipacion;
 	}
